requirements.txt
numpy
numpy-quaternion
pyqtgraph
pyqt5
pyserial
PyOpenGL